backend/requirements.txt
asgiref==3.8.1
Django==5.0.6
djangorestframework==3.15.1
sqlparse==0.5.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
django-debug-toolbar==4.4.2
django-phonenumber-field==7.3.0
phonenumberslite==8.13.38
djangorestframework-simplejwt==5.3.1
pyjwt==2.8.0
